refactoring email group access tests
